requirements.txt
attrs==20.3.0
awsebcli==3.19.2
bcrypt==3.2.0
blessed==1.17.11
boto3==1.16.13
botocore==1.19.13
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.2.1
cycler==0.10.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
eventkit==0.8.6
Flask==1.1.2
future==0.16.0
ib-insync==0.9.64
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.3.1
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
multitasking==0.0.9
nest-asyncio==1.4.2
numpy==1.19.4
pandas==1.1.4
pandas-ta==0.2.23b0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
semantic-version==2.5.0
six==1.15.0
termcolor==1.1.0
texttable==1.6.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
yfinance==0.1.55
